[#264] Moved payment views test into own file
